Rename "pubreg" files to "pubs"
