+++ README.md
@@ -4,7 +4,7 @@ Okdata CLI provides a unified interface to data services for Oslo Origo teams.
 
 * [Install](doc/install.md)
 * [Configuration](doc/configuration.md)
-* [Public services](doc/pubreg.md)
+* [Public services](doc/pubs.md)
 * [Teams](doc/teams.md)
 * [Datasets](doc/datasets.md)
 * [Permissions](doc/permissions.md)
