+++ doc/pubs.md
