--- okdata/cli/commands/pubreg/__init__.py
@@ -1,3 +1,0 @@
-from okdata.cli.commands.pubreg.pubreg import PubsCommand
-
-__all__ = ["PubsCommand"]
