+++ okdata/cli/commands/pubs/__init__.py
@@ -1,0 +1,3 @@
+from okdata.cli.commands.pubs.pubs import PubsCommand
+
+__all__ = ["PubsCommand"]
